Extract enums processing to a separate class
