+++ default/python/stub_generator/stub_generator/__init__.py
@@ -1,0 +1,2 @@
+from stub_generator.stub_generator.argument_name_generator import normalize_name
+from stub_generator.stub_generator.enum_precessor import EnumProcessor
